resolve the warning from EXPECT_EQ(int, enum)

+++ Testing/igtlMessageBaseTest.cxx
@@ -65,7 +65,7 @@ TEST(MessageBaseTest, UNPACKTEST)
 {
   igtl::MessageBase::Pointer messageBaseTest = igtl::MessageBase::New();
   int status = messageBaseTest->Unpack(); // The m_packSize cannot be set, so the unpack cannot be tested
-  EXPECT_EQ(status, messageBaseTest->UNPACK_UNDEF);
+  EXPECT_EQ(status, static_cast<int>(messageBaseTest->UNPACK_UNDEF));
 }
